updates login flow, adds message to redirect uri post successful login
