+++ package.json
@@ -25,7 +25,7 @@
   },
   "icon": "icon/openshift_logo.png",
   "engines": {
-    "vscode": "^1.18.1"
+    "vscode": "^1.12.0"
   },
   "categories": [
     "Other"
